# docs/manekani/apis/cards.openapi.yaml
openapi: 3.0.1
info:
  title: ManeKani API
  description: ManeKani API.
  # termsOfService: http://swagger.io/terms/
  # contact:
  #   name: API Support
  #   url: http://www.swagger.io/support
  #   email: sixels@protonmail.com
  # license:
  #   name: Apache 2.0
  #   url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: "1.0"
servers:
  - url: http://127.0.0.1:8080
...
components:
  schemas:
    common.Response:
      type: object
      required:
        - code
        - data
      properties:
        code:
          type: integer
        data:
          type: object
    common.ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
    subject.Model:
      type: object
      required:
        - id
        - created_at
        - updated_at
        - kind
        - level
        - name
        - slug
        - priority
        - resources
        - study_data
        - dependencies
        - dependents
        - similars
        - deck
        - owner
      properties:
        additional_study_data:
          type: object
        created_at:
          type: string
        deck:
          type: string
          format: uuid
        dependencies:
          type: array
          items:
            type: string
            format: uuid
        dependents:
          type: array
          items:
            type: string
            format: uuid
        id:
          type: string
          format: uuid
        kind:
          type: string
        level:
          type: integer
          format: int32
        name:
          type: string
        owner:
          type: string
        priority:
          type: integer
          format: uint8
        resources:
          type: array
          items:
            $ref: "#/components/schemas/subject.Resource"
        similars:
          type: array
          items:
            type: string
            format: uuid
        slug:
          type: string
        study_data:
          type: array
          items:
            $ref: "#/components/schemas/subject.StudyData"
        updated_at:
          type: string
        value:
          type: string
        value_image:
          type: string
    subject.StudyData:
      type: object
      required:
        - items
        - kind
        - mnemonic
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/subject.StudyItem"
        kind:
          type: string
        mnemonic:
          type: string
    subject.StudyItem:
      type: object
      required:
        - value
        - is_primary
        - is_valid_answer
        - is_hidden
      properties:
        category:
          type: string
        is_hidden:
          type: boolean
        is_primary:
          type: boolean
        is_valid_answer:
          type: boolean
        resource:
          type: string
        value:
          type: string
    subject.Resource:
      type: object
      properties:
        metadata:
          type: object
          additionalProperties:
            type: string
        url:
          type: string
    subject.CreateRequest:
      type: object
      required:
        - deck
        - kind
        - level
        - priority
        - name
        - slug
      properties:
        deck:
          type: string
          format: uuid
          x-oapi-codegen-extra-tags:
            form: "deck"
        dependencies:
          allOf:
            - $ref: "#/components/schemas/subject.IdList"
          x-oapi-codegen-extra-tags:
            form: "dependencies"
        dependents:
          allOf:
            - $ref: "#/components/schemas/subject.IdList"
          x-oapi-codegen-extra-tags:
            form: "dependents"
        kind:
          type: string
          x-oapi-codegen-extra-tags:
            form: "kind"
        level:
          type: integer
          format: int32
          x-oapi-codegen-extra-tags:
            form: "level"
        name:
          type: string
          x-oapi-codegen-extra-tags:
            form: "name"
        priority:
          type: integer
          format: uint8
          x-oapi-codegen-extra-tags:
            form: "priority"
        similar:
          allOf:
            - $ref: "#/components/schemas/subject.IdList"
          x-oapi-codegen-extra-tags:
            form: "similar"
        slug:
          type: string
          x-oapi-codegen-extra-tags:
            form: "slug"
        value:
          type: string
          x-oapi-codegen-extra-tags:
            form: "value"
        value_image:
          type: string
          description: The subject value image
          format: binary
          x-oapi-codegen-extra-tags:
            form: "value_image"
        resources:
          description: The subject resources
          type: array
          items:
            type: string
            format: binary
          x-oapi-codegen-extra-tags:
            form: "resource"
        resources_meta:
          description: Resources metadatas
          type: object
          required:
            - list
          properties:
            list:
              type: array
              items:
                type: object
                additionalProperties:
                  type: string
          x-oapi-codegen-extra-tags:
            form: "resources_meta"
        study_data:
          type: object
          required:
            - list
          properties:
            list:
              type: array
              items:
                $ref: "#/components/schemas/subject.StudyData"
          x-oapi-codegen-extra-tags:
            form: "study_data"
        additional_study_data:
          type: object
          x-oapi-codegen-extra-tags:
            form: "additional_study_data"
    subject.CreateResponse:
      $ref: "#/components/schemas/subject.Model"
...
    subject.GetAllResponse:
      type: array
      items:
        type: object
        required:
          - id
          - kind
          - level
          - name
          - slug
          - priority
          - study_data
          - dependencies
          - dependents
          - similars
          - deck
          - owner
        properties:
          deck:
            type: string
            format: uuid
          dependencies:
            type: array
            items:
              type: string
              format: uuid
          dependents:
            type: array
            items:
              type: string
              format: uuid
          id:
            type: string
            format: uuid
          kind:
            type: string
          level:
            type: integer
            format: int32
          name:
            type: string
          owner:
            type: string
          priority:
            type: integer
            format: uint8
          similars:
            type: array
            items:
              type: string
              format: uuid
          slug:
            type: string
          study_data:
            type: array
            items:
              $ref: "#/components/schemas/subject.StudyData"
          value:
            type: string
          value_image:
            type: string
...
    subject.IdList:
      type: object
      required:
        - list
      properties:
        list:
          type: array
          items:
            type: string
            format: uuid
...
  securitySchemes:
    ApiKey:
      type: apiKey
      description: API key authentication
      name: Authorization
      in: header
    Login:
      type: apiKey
      description: Login at http://127.0.0.1:4455/login and copy the contents of the 'ory_kratos_session' cookie
      name: ory_kratos_session
      in: cookie
paths:
  /api/v1/subjects:
    get:
      tags:
        - subject
      summary: Query all subjects
      description: Return a list of all subjects
      operationId: get-subjects
      parameters:
        - name: decks
          in: query
          schema:
            type: string
        - name: ids
          in: query
          schema:
            type: string
        - name: kinds
          in: query
          schema:
            type: string
        - name: levels
          in: query
          schema:
            type: string
        - name: owners
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: slugs
          in: query
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/common.Response"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/subject.GetAllResponse"
    post:
      tags:
        - subject
      summary: Create a new subject
      description: Creates a subject with the given values
      operationId: create-subject
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/subject.CreateRequest"
        required: true
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/common.Response"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/subject.CreateResponse"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/common.ErrorResponse"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/common.ErrorResponse"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/common.ErrorResponse"
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/common.ErrorResponse"
      security:
        - ApiKey: ["subject:create"]
        - Login: []
...